Create Scrabble
In this project I start with a common dict key:value with a score value associated with each alphabet letter.

Then, I calculate the total score for the word 'BROWNIE' based on the dictionary above.

Finally, I calculate the total score of the words available in the value of the key:value pair of the 'player_to_words' dictionary. 

Enjoy!